+++ pkg/querier/blocks_store_queryable_test.go
@@ -2688,6 +2688,10 @@ func (m *storeGatewayClientMock) LabelValues(_ context.Context, r *storepb.Label
 	return m.mockedLabelValuesResponse, m.mockedLabelValuesErr
 }
 
+func (m *storeGatewayClientMock) Cardinality(_ context.Context, _ *storegatewaypb.CardinalityRequest, _ ...grpc.CallOption) (*storegatewaypb.CardinalityResponse, error) {
+	return &storegatewaypb.CardinalityResponse{}, nil
+}
+
 func (m *storeGatewayClientMock) RemoteAddress() string {
 	return m.remoteAddr
 }